+++ src/helpers/isValidUrl.ts
@@ -1,6 +1,6 @@
 import { URL_REGEX } from '@/constants/regexp.js';
 
-export function isValidUrl(url?: string): url is string {
+export function isValidUrl(url: string): boolean {
     if (!url) return false;
     URL_REGEX.lastIndex = 0;
     return URL_REGEX.test(url);
