--- src/helpers/resolveAvatarFallbackUrl.ts
@@ -1,4 +1,0 @@
-export function resolveAvatarFallbackUrl(url: string | undefined, isDarkMode = false) {
-    if (!url?.startsWith('https://cdn.stamp.fyi/avatar/eth:')) return url;
-    return isDarkMode ? '/image/firefly-dark-avatar.png' : '/image/firefly-light-avatar.png';
-}
