+++ src/helpers/resolveTCOLink.ts
@@ -17,5 +17,6 @@ async function resolver(u: string): Promise<string | null> {
     if (url) cache.set(u, url);
     return url ?? null;
 }
+
 /** Resolve a https://t.co/ link to it's real address. */
 export const resolveTCOLink = memoizePromise(memoize, resolver, (x) => x);
