+++ src/locales/zh-Hant/messages.po
@@ -1799,4 +1799,3 @@ msgstr ""
 
 msgid "<0/> quoted your <1><2/></1>"
 msgstr ""
-
